\name{PH_examp}
\docType{data}
\alias{PH_examp}
\title{Data example for proportional hazard model.}
\description{
This data set is randomly generated from the proportional hazards transformation model as an example.
}
\usage{data(PH_examp)}
\format{A data frame containing survival time, right censoring indicator, predictors gender and age, and a trivial variable called error.} 
\keyword{datasets}