\name{PO_examp}
\docType{data}
\alias{PO_examp}
\title{Data example for proportional odds model.}
\description{
This data set is randomly generated from the proportional odds transformation model as an example.
}
\usage{data(PO_examp)}
\format{A data frame containing survival time, right censoring indicator, predictors gender and age, and a trivial variable called error.} 
\keyword{datasets}