\name{print.summary.TransModel}
\alias{print.summary.TransModel}

\title{
Print method for the summary output of TransModel.
}
\description{
A summary table with coefficient estimates, standard error and p-values will be printed.
}
\usage{
\method{print}{summary.TransModel}(x, ...)
}
\arguments{
  \item{x}{
An object from the function summary.TransModel.
}
  \item{\dots}{
Other print arguments.
}
}



