\name{print.TransModel}
\alias{print.TransModel}
\title{
Print method for the function TransMod.
}
\description{
Estimated coefficients and covariance matrix will be printed for the fitted linear transformation model
 from TransMod.
}
\usage{
\method{print}{TransModel}(x, ...)
}
\arguments{
  \item{x}{
An object from the function TransModel.
}
  \item{\dots}{
Other auguments to be specified.
}
}
