\name{summary.TransModel}
\alias{summary.TransModel}
\title{
Summary of hypothesis tests for the coefficients in the linear transformation model.
}
\description{
Estimates, standard errors, test statistics and p values are presented for each coefficient in the linear
 transformation model.}
\usage{
\method{summary}{TransModel}(object, ...)
}
\arguments{
  \item{object}{
An object from the function TransModel.
}
  \item{\dots}{
Other auguments to be specified.
}
}
