\name{TransModel-package}
\alias{TransModel-package}
\docType{package}
\title{
\packageTitle{TransModel}
}
\description{
Semiparametric analysis for linear transformation model, such as the proportional hazards model
 and the proportional odds model, is fitted based on a unified method proposed in Chen(2002) for right censored survival data.
}
\details{
The main function TransModel gives estimates for coefficients and covariance matrix for the predictors
in the linear transformation model. Print and summary method can be applied to the returned object.
}
\author{
\packageAuthor{TransModel}

Maintainer: \packageMaintainer{TransModel}
}
\references{
Kani Chen, et al., Semiparametric analysis of transformation models with censored data. Biometrika, 
89(3), 659-668, 2002.
}

